pyproject.toml
[tool.poetry]
name = "my-flasktours"
version = "0.1.0"
description = ""
authors = ["rodion228235 <rodionkovalenko201@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
